refactor: flattern packages internal
