+++ account/account.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/kkgo-software-engineering/workshop/config"
-	"github.com/kkgo-software-engineering/workshop/middleware/mlog"
+	"github.com/kkgo-software-engineering/workshop/mlog"
 	"github.com/labstack/echo/v4"
 	"go.uber.org/zap"
 )
