+++ middleware/auth_test.go
@@ -1,4 +1,4 @@
-package auth
+package middleware
 
 import (
 	"encoding/base64"
